+++ Kernel/Modules/AgentTicketBulk.pm
@@ -67,6 +67,61 @@ sub new {
 sub Run {
     my ( $Self, %Param ) = @_;
 
+    if ( $Self->{Subaction} eq 'CancelAndUnlockTickets' ) {
+
+        my @TicketIDs
+            = grep {$_}
+            $Self->{ParamObject}->GetArray( Param => 'LockedTicketID' );
+
+        # challenge token check for write action
+        $Self->{LayoutObject}->ChallengeTokenCheck();
+
+        # check needed stuff
+        if ( !@TicketIDs ) {
+            return $Self->{LayoutObject}->ErrorScreen(
+                Message => 'Can\'t lock Tickets, no TicketIDs are given!',
+                Comment => 'Please contact the admin.',
+            );
+        }
+
+        my $Message = '';
+
+        TICKET_ID:
+        for my $TicketID (@TicketIDs) {
+
+            my $Access = $Self->{TicketObject}->TicketPermission(
+                Type     => 'lock',
+                TicketID => $TicketID,
+                UserID   => $Self->{UserID}
+            );
+
+            # error screen, don't show ticket
+            if ( !$Access ) {
+                return $Self->{LayoutObject}->NoPermission( WithHeader => 'yes' );
+            }
+
+            # set unlock
+            my $Lock = $Self->{TicketObject}->TicketLockSet(
+                TicketID => $TicketID,
+                Lock     => 'unlock',
+                UserID   => $Self->{UserID},
+            );
+            if ( !$Lock ) {
+                $Message .= "$TicketID,";
+            }
+        }
+
+        if ( $Message ne '' ) {
+            return $Self->{LayoutObject}
+                ->ErrorScreen( Message => "Ticket ($Message) is not unlocked!", );
+        }
+
+        return $Self->{LayoutObject}->Redirect(
+            OP => $Self->{LastScreenOverview},
+        );
+
+    }
+
     # check if bulk feature is enabled
     if ( !$Self->{ConfigObject}->Get('Ticket::Frontend::BulkFeature') ) {
         return $Self->{LayoutObject}->ErrorScreen(
@@ -227,8 +282,10 @@ sub Run {
 
     # process tickets
     my @TicketIDSelected;
-    my $ActionFlag = 0;
-    my $Counter    = 1;
+    my $LockedTickets = '';
+    my $ActionFlag    = 0;
+    my $Counter       = 1;
+    $Param{TicketsWereLocked} = 0;
 
     TICKET_ID:
     for my $TicketID (@TicketIDs) {
@@ -254,8 +311,6 @@ sub Run {
             next TICKET_ID;
         }
 
-        $Param{TicketsWereLocked} = 0;
-
         # check if it's already locked by somebody else
         if ( !$Self->{Config}->{RequiredLock} ) {
             $Output .= $Self->{LayoutObject}->Notify(
@@ -279,6 +334,7 @@ sub Run {
                 }
             }
             else {
+                $LockedTickets .= "LockedTicketID=" . $TicketID . ';';
                 $Param{TicketsWereLocked} = 1;
             }
 
@@ -673,8 +729,9 @@ sub Run {
         %Param,
         %GetParam,
         %Time,
-        TicketIDs => \@TicketIDSelected,
-        Errors    => \%Error,
+        TicketIDs     => \@TicketIDSelected,
+        LockedTickets => $LockedTickets,
+        Errors        => \%Error,
     );
     $Output .= $Self->{LayoutObject}->Footer(
         Type => 'Small',
@@ -1016,6 +1073,20 @@ sub _Mask {
             Data => {
                 URL => $Self->{LastScreenOverview},
                 }
+        );
+
+        # show undo&close link
+        $Self->{LayoutObject}->Block(
+            Name => 'UndoClosePopup',
+            Data => {%Param},
+        );
+    }
+    else {
+
+        # show cancel&close link
+        $Self->{LayoutObject}->Block(
+            Name => 'CancelClosePopup',
+            Data => {%Param},
         );
     }